+++ src/assets/browser/app/src/controller/global/mod.rs
@@ -2,10 +2,8 @@ mod booting;
 mod global;
 mod apprunner;
 mod app;
-mod scheduler;
 
 pub use self::booting::Booting;
 pub use self::global::{ Global, GlobalWeak, setup_global };
 pub use self::apprunner::{ AppRunner, AppRunnerWeak };
 pub use self::app::App;
-pub use self::scheduler::{ Scheduler, SchedRun, SchedulerGroup };
