+++ src/assets/browser/app/src/controller/mod.rs
@@ -1,3 +1,4 @@
 pub mod global;
 pub mod input;
 pub mod output;
+pub mod scheduler;
