+++ src/assets/browser/app/src/controller/output/mod.rs
@@ -1,4 +1,3 @@
-mod jank;
 mod outputaction;
 mod report;
 mod viewportreport;
@@ -6,4 +5,3 @@ mod viewportreport;
 pub use self::outputaction::OutputAction;
 pub use self::report::Report;
 pub use self::viewportreport::ViewportReport;
-pub use self::jank::JankBuster;
